scraper.requirements.txt
redis==4.5.1
selenium==4.15.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.28.1